+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Added
 - UFG skip list functionality. [Trello 1974](https://trello.com/c/44xq8dze)
 - New DOM scripts features. [Trello 2268](https://trello.com/c/x6ImzMue)
+### Fixed
+- Debug resource writer is now working. [Trello 2310](https://trello.com/c/EOSqNaoG)
 
 ## [3.183.0] - 2020-11-09
 ### Updated